Updated from Raid Guild latest developments

+++ src/math/Sqrt.sol
@@ -5,13 +5,12 @@ pragma solidity ^0.8.0;
 // a library for performing sqrt
 
 library Sqrt {
-
     // babylonian method (https://en.wikipedia.org/wiki/Methods_of_computing_square_roots#Babylonian_method)
     function sqrt(uint256 y) internal pure returns (uint256) {
         uint256 z;
         if (y > 3) {
             z = y;
-            uint x = y / 2 + 1;
+            uint256 x = y / 2 + 1;
             while (x < z) {
                 z = x;
                 x = (y / x + x) / 2;
